fabtests/shm,sm2: filter memory types at collection by device availability
Parametrize memory_type at collection from a candidate list, dropping
cuda permutations whose device is absent on the owning endpoint so they
are never collected. Fall back to all candidates when server/client ids
are absent or detection fails.

The shared client_server_have_device function is moved to common.py
