+++ fabtests/pytest/efa/conftest.py
@@ -4,6 +4,7 @@ import subprocess
 import time
 from retrying import retry
 from common import (
+    client_server_have_device,
     num_hmem_devices,
     test_selected_by_marker,
     has_ssh_connection_err_msg,
@@ -101,25 +102,6 @@ def add_fabric_and_message_size_parametrization(metafunc, fabric_marker, sizes_m
             f"no kwarg naming {test_type!r} (have {sorted(sizes_marker.kwargs)})"
         )
     metafunc.parametrize("message_sizes", sizes)
-
-
-def client_server_have_device(memory_token, server_id, client_id):
-    """
-    Return True if both client and server endpoints named in a memory-type
-    token have the hmem device that token requires.
-
-    Any SSH/detection failure propagates to the caller,
-    which falls back to including all candidate memory types.
-    """
-    client_memory_type, server_memory_type = memory_token.split("_to_")
-    for memory_type_name, ip in ((client_memory_type, client_id),
-                                 (server_memory_type, server_id)):
-        if memory_type_name == "host":
-            # host memory needs no accelerator device
-            continue
-        if num_hmem_devices(ip, memory_type_name) <= 0:
-            return False
-    return True
 
 
 def add_memory_type_parametrization(metafunc, memory_type_marker):
